docs: move docs into their own folder (#10)
