+++ src/docs/views/docs-versions.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
 
-import pkg from "../../package.json";
+import pkg from "../../../package.json";
 
 interface Props {
   versions: string[];
